// models/enums.ts
export enum BookingStatus {
  PENDING = "PENDING",
  BOOKED = "BOOKED",
  RUNNING = "RUNNING",
  CLOSED = "CLOSED",
  CANCELLED = "CANCELLED",
  REJECTED = "REJECTED",
  SELECTED = "SELECTED",
  AVAILABLE = "AVAILABLE",
}
export enum slotType {
  PENDING = "PENDING",
  BOOKED = "BOOKED",
  RUNNING = "RUNNING",
  CLOSED = "CLOSED",
  CANCELLED = "CANCELLED",
  REJECTED = "REJECTED",
}

export enum DiscountType {
  Amount = "AMT",
  Percent = "PER",
}